playlist-downloader/requirements.txt
# Core async I/O
aiofiles==25.1.0

# System resource detection
psutil==7.0.0

# Configuration management
python-dotenv==1.2.1

# Rich terminal UI
rich==14.2.0

# Spotify API client
spotipy==2.25.1

# Progress bars
tqdm==4.67.1

# Required for yt-dlp (installed separately)
# yt-dlp - Install with: pip install yt-dlp
# or: python -m pip install yt-dlp
